changed long back to float

+++ charades_dataset.py
@@ -121,7 +121,7 @@ class Charades(data_utl.Dataset):
         imgs = self.transforms(imgs)
         # print(imgs.shape)
 
-        return video_to_tensor(imgs), torch.from_numpy(label).long()
+        return video_to_tensor(imgs), torch.from_numpy(label)
 
     def __len__(self):
         return len(self.data)
